requirements.txt
Flask>=2.0
flask-bcrypt
Flask-SQLAlchemy
flask-cors
PyJWT
requests
# Optional (only needed if you use Postgres)
psycopg2-binary
